use integer

+++ com.unity.toonshader/Runtime/Shaders/URP/ToonURP_3Das2D.shader
@@ -60,9 +60,9 @@ Shader "Toon/Toon 3D as 2D (URP)"{
         // Put this in the last bit of our stencil value for maximum compatibility with sprite mask
         [IntRange] _StencilRef ("Stencil Reference", Range(0, 255)) = 128
         
-        [Enum(UnityEngine.Rendering.CompareFunction)] _StencilComp ("Stencil Compare", Float) = 8   // Always
-        [Enum(UnityEngine.Rendering.StencilOp)] _StencilOpPass ("Stencil Pass Op", Float) = 2       // Replace
-        [Enum(UnityEngine.Rendering.StencilOp)] _StencilOpFail ("Stencil Fail Op", Float) = 0       // Keep
+        [Enum(UnityEngine.Rendering.CompareFunction)] _StencilComp ("Stencil Compare", Integer) = 8   // Always
+        [Enum(UnityEngine.Rendering.StencilOp)] _StencilOpPass ("Stencil Pass Op", Integer) = 2       // Replace
+        [Enum(UnityEngine.Rendering.StencilOp)] _StencilOpFail ("Stencil Fail Op", Integer) = 0       // Keep
 
     }
